feat: split chat page components and state management
