+++ lib/features/chat/providers/chat_provider.dart
@@ -1,9 +1,11 @@
-import 'dart:math';
-
-import 'package:dio/dio.dart';
 import 'package:flutter/foundation.dart';
+import 'package:servllama/features/chat/controllers/chat_conversation_controller.dart';
+import 'package:servllama/features/chat/controllers/chat_generation_controller.dart';
+import 'package:servllama/features/chat/controllers/chat_id_generator.dart';
+import 'package:servllama/features/chat/controllers/chat_model_controller.dart';
+import 'package:servllama/features/chat/controllers/chat_session_list_controller.dart';
+import 'package:servllama/features/chat/controllers/streaming_chat_message_notifier.dart';
 import 'package:servllama/features/chat/models/chat_message_record.dart';
-import 'package:servllama/features/chat/models/chat_message_version_record.dart';
 import 'package:servllama/features/chat/models/chat_model_option.dart';
 import 'package:servllama/features/chat/models/chat_session_record.dart';
 import 'package:servllama/features/chat/repositories/chat_session_repository.dart';
@@ -14,1245 +16,272 @@ class ChatProvider extends ChangeNotifier {
     ChatSessionRepository? repository,
     LlamaChatApiClient? apiClient,
   }) : _repository = repository ?? ChatSessionRepository(),
-       _apiClient = apiClient ?? LlamaChatApiClient();
+       _apiClient = apiClient ?? LlamaChatApiClient() {
+    _idGenerator = ChatIdGenerator();
+    _sessionList = ChatSessionListController(repository: _repository);
+    _conversation = ChatConversationController(
+      repository: _repository,
+      sessionList: _sessionList,
+      messageWindowSize: messageWindowSize,
+    );
+    _models = ChatModelController(apiClient: _apiClient);
+    _generation = ChatGenerationController(
+      repository: _repository,
+      apiClient: _apiClient,
+      sessionList: _sessionList,
+      conversation: _conversation,
+      models: _models,
+      idGenerator: _idGenerator,
+      defaultSessionTitle: defaultSessionTitle,
+    );
+
+    _sessionList.addListener(_notify);
+    _conversation.addListener(_notify);
+    _models.addListener(_notify);
+    _generation.addListener(_notify);
+  }
 
   static const String defaultSessionTitle = '新会话';
   static const int messageWindowSize = 30;
 
   final ChatSessionRepository _repository;
   final LlamaChatApiClient _apiClient;
-  final Random _random = Random();
 
-  List<ChatSessionRecord> _sessions = <ChatSessionRecord>[];
-  List<ChatMessageRecord> _visibleMessages = <ChatMessageRecord>[];
-  List<ChatModelOption> _models = <ChatModelOption>[];
-
-  bool _isInitialized = false;
-  bool _isLoading = false;
-  bool _isLoadingMessages = false;
-  bool _isLoadingOlderMessages = false;
-  bool _isRefreshingModels = false;
-  bool _isServerRunning = false;
-  String _baseUrl = 'http://127.0.0.1:8080';
-  String? _selectedSessionId;
-  String? _currentModelId;
-  String? _loadingModelId;
-  String _sessionQuery = '';
-  bool _isSending = false;
-  ChatMessageRecord? _draftAssistantMessage;
-  CancelToken? _activeCancelToken;
-  String? _lastErrorMessage;
-  List<String> _pendingImageAttachments = [];
-  int _sessionLoadGeneration = 0;
+  late final ChatIdGenerator _idGenerator;
+  late final ChatSessionListController _sessionList;
+  late final ChatConversationController _conversation;
+  late final ChatModelController _models;
+  late final ChatGenerationController _generation;
 
   @visibleForTesting
   LlamaChatApiClient get apiClient => _apiClient;
 
-  List<ChatSessionRecord> get sessions =>
-      List<ChatSessionRecord>.unmodifiable(_sessions);
-  List<ChatModelOption> get models =>
-      List<ChatModelOption>.unmodifiable(_models);
+  StreamingChatMessageNotifier get streamingMessages =>
+      _generation.streamingMessages;
 
-  bool get isLoading => _isLoading;
-  bool get isLoadingMessages => _isLoadingMessages;
-  bool get isLoadingOlderMessages => _isLoadingOlderMessages;
-  bool get isRefreshingModels => _isRefreshingModels;
-  bool get isServerRunning => _isServerRunning;
-  bool get isSending => _isSending;
-  String get sessionQuery => _sessionQuery;
-  String? get loadingModelId => _loadingModelId;
-  String? get currentModelId => _currentModelId;
-  String? get draftMessageId => _draftAssistantMessage?.id;
-  String? get lastErrorMessage => _lastErrorMessage;
+  List<ChatSessionRecord> get sessions => _sessionList.sessions;
+  List<ChatModelOption> get models => _models.models;
+
+  bool get isLoading => _sessionList.isLoading;
+  bool get isLoadingMessages => _conversation.isLoadingMessages;
+  bool get isLoadingOlderMessages => _conversation.isLoadingOlderMessages;
+  bool get isRefreshingModels => _models.isRefreshingModels;
+  bool get isServerRunning => _models.isServerRunning;
+  bool get isSending => _generation.isSending;
+  String get sessionQuery => _sessionList.query;
+  String? get loadingModelId => _models.loadingModelId;
+  String? get currentModelId => _models.currentModelId;
+  String? get draftMessageId => _generation.draftMessageId;
+  String? get lastErrorMessage => _generation.lastErrorMessage;
   List<String> get pendingImageAttachments =>
-      List<String>.unmodifiable(_pendingImageAttachments);
+      _generation.pendingImageAttachments;
 
-  bool get canManageSessions => !_isSending && _loadingModelId == null;
-  bool get canSelectModels =>
-      _isServerRunning && !_isSending && _loadingModelId == null;
-  bool get canSend =>
-      !_isSending &&
-      _loadingModelId == null &&
-      _isServerRunning &&
-      currentModel?.isLoaded == true;
-  bool get canManageMessages => !_isSending && _loadingModelId == null;
+  bool get canManageSessions => _generation.canManageSessions;
+  bool get canSelectModels => _generation.canSelectModels;
+  bool get canSend => _generation.canSend;
+  bool get canManageMessages => _generation.canManageMessages;
 
-  ChatSessionRecord? get selectedSession => _findSession(_selectedSessionId);
-  bool get isShowingDraftSession => _selectedSessionId == null;
+  ChatSessionRecord? get selectedSession => _conversation.selectedSession;
+  bool get isShowingDraftSession => _conversation.isShowingDraftSession;
   String get currentSessionTitle =>
       selectedSession?.title ?? defaultSessionTitle;
 
-  bool get hasOlderMessages {
-    final session = selectedSession;
-    if (session == null || _visibleMessages.isEmpty) {
-      return false;
-    }
-    final firstVisibleId = _visibleMessages.first.id;
-    final firstVisibleIndex = session.messageIds.indexOf(firstVisibleId);
-    return firstVisibleIndex > 0;
-  }
+  bool get hasOlderMessages => _conversation.hasOlderMessages;
 
-  ChatModelOption? get currentModel => _findModel(_currentModelId);
-
-  String get currentModelDisplayName {
-    final model = currentModel;
-    if (model != null) {
-      return model.displayName;
-    }
-    if (_currentModelId != null && _currentModelId!.trim().isNotEmpty) {
-      return _currentModelId!;
-    }
-    return '未选择模型';
-  }
-
-  String get modelSelectorLabel {
-    final model = currentModel;
-    if (model != null && model.isLoaded) {
-      return model.displayName;
-    }
-    return '选择模型';
-  }
-
-  String get currentModelStatusLabel {
-    final model = currentModel;
-    if (model == null) {
-      return _currentModelId == null ? '请选择模型' : '模型暂不可用';
-    }
-    return _modelStatusLabel(model.status);
-  }
-
-  List<ChatModelOption> get loadedModels =>
-      _models.where((model) => model.isLoaded).toList(growable: false);
-
-  List<ChatModelOption> get availableModels =>
-      _models.where((model) => !model.isLoaded).toList(growable: false);
-
-  List<ChatMessageRecord> get visibleMessages {
-    return List<ChatMessageRecord>.from(_visibleMessages, growable: false);
-  }
-
-  List<ChatSessionRecord> get filteredSessions {
-    final normalizedQuery = _normalizedSessionQuery;
-    if (normalizedQuery.isEmpty) {
-      return sessions;
-    }
-    return _sessions
-        .where(
-          (session) => session.title.toLowerCase().contains(normalizedQuery),
-        )
-        .toList(growable: false);
-  }
-
-  String get inputHintText {
-    if (!_isServerRunning) {
-      return '请先启动服务器';
-    }
-    if (_loadingModelId != null) {
-      return '模型加载中...';
-    }
-    if (_currentModelId == null) {
-      return '请先选择模型';
-    }
-    if (currentModel?.isLoaded != true) {
-      return '当前模型未加载';
-    }
-    return '输入消息';
-  }
+  ChatModelOption? get currentModel => _models.currentModel;
+  String get currentModelDisplayName => _models.currentModelDisplayName;
+  String get modelSelectorLabel => _models.modelSelectorLabel;
+  String get currentModelStatusLabel => _models.currentModelStatusLabel;
+  List<ChatModelOption> get loadedModels => _models.loadedModels;
+  List<ChatModelOption> get availableModels => _models.availableModels;
+  List<ChatMessageRecord> get visibleMessages => _conversation.visibleMessages;
+  List<ChatSessionRecord> get filteredSessions => _sessionList.filteredSessions;
+  String get inputHintText => _models.inputHintText;
 
   Future<void> load() async {
-    if (_isInitialized || _isLoading) {
-      return;
-    }
-
-    _isLoading = true;
-    notifyListeners();
-
-    try {
-      _sessions = List<ChatSessionRecord>.from(
-        await _repository.loadSessions(),
-      );
-      if (_findSession(_selectedSessionId) == null) {
-        _selectedSessionId = null;
-        _visibleMessages = <ChatMessageRecord>[];
-      }
-      _isInitialized = true;
-    } catch (_) {
-    } finally {
-      _isLoading = false;
-      notifyListeners();
-    }
+    await _sessionList.load();
+    _conversation.clearIfSelectedSessionMissing(notify: false);
   }
 
   void updateSessionQuery(String value) {
-    final nextQuery = value;
-    if (_sessionQuery == nextQuery) {
-      return;
-    }
-    _sessionQuery = nextQuery;
-    notifyListeners();
+    _sessionList.updateQuery(value);
   }
 
   void updateServerState({
     required String baseUrl,
     required bool isServerRunning,
   }) {
-    var changed = false;
-
-    if (_baseUrl != baseUrl) {
-      _baseUrl = baseUrl;
-      _apiClient.updateBaseUrl(baseUrl);
-    }
-
-    final stopped = _isServerRunning && !isServerRunning;
-    if (_isServerRunning != isServerRunning) {
-      _isServerRunning = isServerRunning;
-      changed = true;
-    }
-
+    final stopped = _models.updateServerState(
+      baseUrl: baseUrl,
+      isServerRunning: isServerRunning,
+    );
     if (stopped) {
-      _models = <ChatModelOption>[];
-      _loadingModelId = null;
-      _activeCancelToken?.cancel('server stopped');
-      changed = true;
-    }
-
-    if (changed) {
-      notifyListeners();
+      _generation.cancelForServerStop();
     }
   }
 
   void updateChatTimeout(Duration timeout) {
-    _apiClient.updateReceiveTimeout(timeout);
+    _models.updateChatTimeout(timeout);
   }
 
   Future<void> refreshModels() async {
-    if (_isRefreshingModels || !_isServerRunning) {
-      return;
-    }
-
-    _isRefreshingModels = true;
-    notifyListeners();
-
-    try {
-      _models = await _apiClient.fetchModels();
-      final selectedModel = currentModel;
-      if (_currentModelId != null &&
-          (selectedModel == null || !selectedModel.isLoaded)) {
-        _currentModelId = null;
-      }
-    } catch (_) {
-    } finally {
-      _isRefreshingModels = false;
-      notifyListeners();
-    }
+    await _models.refreshModels();
   }
 
   Future<void> createSession() async {
-    if (!canManageSessions) {
+    if (!canManageSessions || _conversation.selectedSessionId == null) {
       return;
     }
-
-    if (_selectedSessionId == null) {
-      return;
-    }
-
-    _selectedSessionId = null;
-    _visibleMessages = <ChatMessageRecord>[];
-    _isLoadingMessages = false;
-    _isLoadingOlderMessages = false;
-    _sessionLoadGeneration += 1;
-    notifyListeners();
+    _conversation.clearSelection();
   }
 
   Future<void> renameSession(String sessionId, String title) async {
     if (!canManageSessions) {
       return;
     }
-
-    final normalizedTitle = title.trim();
-    if (normalizedTitle.isEmpty) {
-      return;
-    }
-
-    final session = _findSession(sessionId);
-    if (session == null) {
-      return;
-    }
-
-    final updated = session.copyWith(
-      title: normalizedTitle,
-      updatedAt: DateTime.now(),
-    );
-    await _saveSessionLocally(updated);
+    await _sessionList.renameSession(sessionId, title);
   }
 
   Future<void> deleteSession(String sessionId) async {
     if (!canManageSessions) {
       return;
     }
-
-    final session = _findSession(sessionId);
-    if (session == null) {
-      return;
+    final isSelected = _conversation.selectedSessionId == sessionId;
+    if (isSelected) {
+      _conversation.clearSelection(notify: false);
     }
-
-    final nextSessions = List<ChatSessionRecord>.from(_sessions);
-    nextSessions.removeWhere((item) => item.id == sessionId);
-    _sessions = nextSessions;
-    await _repository.deleteSession(sessionId);
-
-    if (_selectedSessionId == sessionId) {
-      _selectedSessionId = null;
-      _visibleMessages = <ChatMessageRecord>[];
-      _isLoadingMessages = false;
-      _isLoadingOlderMessages = false;
-      _sessionLoadGeneration += 1;
+    final deleted = await _sessionList.deleteSession(sessionId);
+    if (isSelected && !deleted) {
+      _notify();
     }
-    notifyListeners();
   }
 
-  bool canRegenerateMessage(String messageId) {
-    if (!canSend) {
+  bool beginSessionSelection(
+    String sessionId, {
+    bool keepVisibleMessages = true,
+  }) {
+    if (!canManageSessions) {
       return false;
     }
-    final targetIndex = _messageIndex(_visibleMessages, messageId);
-    if (targetIndex < 0) {
-      return false;
-    }
-    final targetMessage = _visibleMessages[targetIndex];
-    if (targetMessage.role == ChatRole.user) {
-      return true;
-    }
-    final sessionMessageIds = selectedSession?.messageIds ?? const <String>[];
-    final firstSessionMessageId = sessionMessageIds.isEmpty
-        ? null
-        : sessionMessageIds.first;
-    return _nearestUserMessageIndex(_visibleMessages, targetIndex) >= 0 ||
-        firstSessionMessageId != targetMessage.id;
-  }
-
-  Future<void> editMessage({
-    required String messageId,
-    required String newContent,
-  }) async {
-    if (!canManageMessages) {
-      return;
-    }
-
-    final session = selectedSession;
-    if (session == null) {
-      return;
-    }
-
-    final normalizedContent = newContent.trim();
-    if (normalizedContent.isEmpty) {
-      return;
-    }
-
-    final messages = await _repository.loadAllMessages(session);
-    final messageIndex = _messageIndex(messages, messageId);
-    if (messageIndex < 0) {
-      return;
-    }
-
-    final currentMessage = messages[messageIndex];
-    if (currentMessage.content == normalizedContent) {
-      return;
-    }
-
-    final updatedMessages = List<ChatMessageRecord>.from(messages);
-    final updatedMessage = currentMessage.copyWith(content: normalizedContent);
-    updatedMessages[messageIndex] = updatedMessage;
-
-    if (currentMessage.versionIds.isNotEmpty &&
-        currentMessage.role == ChatRole.assistant) {
-      final versionId = _selectedVersionId(currentMessage);
-      if (versionId != null) {
-        final version = await _repository.loadMessageVersion(versionId);
-        if (version != null) {
-          await _repository.saveMessageVersion(
-            ChatMessageVersionRecord(
-              id: version.id,
-              messageId: version.messageId,
-              content: normalizedContent,
-              createdAt: version.createdAt,
-              modelName: version.modelName,
-              reasoningContent: version.reasoningContent,
-              imageFilePaths: version.imageFilePaths,
-            ),
-          );
-        }
-      }
-    }
-
-    await _saveSessionWithMessagesLocally(
-      session.copyWith(messages: updatedMessages, updatedAt: DateTime.now()),
-      updatedMessages,
+    return _conversation.beginSessionSelection(
+      sessionId,
+      keepVisibleMessages: keepVisibleMessages,
     );
   }
 
-  Future<void> deleteMessage(String messageId) async {
-    if (!canManageMessages) {
+  Future<void> loadSelectedSessionMessages(String sessionId) async {
+    if (!canManageSessions) {
       return;
     }
-
-    final session = selectedSession;
-    if (session == null) {
-      return;
-    }
-
-    final messages = await _repository.loadAllMessages(session);
-    final messageIndex = _messageIndex(messages, messageId);
-    if (messageIndex < 0) {
-      return;
-    }
-
-    final targetMessage = messages[messageIndex];
-    final updatedMessages = List<ChatMessageRecord>.from(messages)
-      ..removeAt(messageIndex);
-    await _repository.deleteMessages(<ChatMessageRecord>[targetMessage]);
-    await _saveSessionWithMessagesLocally(
-      session.copyWith(messages: updatedMessages, updatedAt: DateTime.now()),
-      updatedMessages,
-    );
-  }
-
-  Future<void> selectMessageVersion({
-    required String messageId,
-    required int versionIndex,
-  }) async {
-    if (!canManageMessages) {
-      return;
-    }
-
-    final session = selectedSession;
-    if (session == null) {
-      return;
-    }
-
-    final messages = await _repository.loadAllMessages(session);
-    final messageIndex = _messageIndex(messages, messageId);
-    if (messageIndex < 0) {
-      return;
-    }
-
-    final message = messages[messageIndex];
-    if (message.role != ChatRole.assistant ||
-        versionIndex < 0 ||
-        versionIndex >= message.versionIds.length ||
-        versionIndex == message.currentVersionIndex) {
-      return;
-    }
-
-    final version = await _repository.loadMessageVersion(
-      message.versionIds[versionIndex],
-    );
-    if (version == null) {
-      return;
-    }
-
-    final updatedMessages = List<ChatMessageRecord>.from(messages);
-    updatedMessages[messageIndex] = _messageWithVersion(
-      message,
-      version,
-      versionIndex,
-    );
-
-    await _saveSessionWithMessagesLocally(
-      session.copyWith(messages: updatedMessages, updatedAt: DateTime.now()),
-      updatedMessages,
-    );
-  }
-
-  Future<void> regenerateFromMessage(String messageId) async {
-    if (!canRegenerateMessage(messageId)) {
-      return;
-    }
-
-    final session = selectedSession;
-    final model = currentModel;
-    if (session == null || model == null) {
-      return;
-    }
-
-    final messages = await _repository.loadAllMessages(session);
-    final targetIndex = _messageIndex(messages, messageId);
-    if (targetIndex < 0) {
-      return;
-    }
-    final userIndex = _nearestUserMessageIndex(messages, targetIndex);
-    if (userIndex < 0) {
-      return;
-    }
-
-    final targetMessage = messages[targetIndex];
-    if (targetMessage.role == ChatRole.assistant) {
-      await _regenerateAssistantMessageVersion(
-        session: session,
-        messages: messages,
-        targetIndex: targetIndex,
-        model: model,
-      );
-      return;
-    }
-
-    final retainedMessages = List<ChatMessageRecord>.from(
-      messages.take(userIndex + 1),
-    );
-    final removedMessages = messages.skip(userIndex + 1);
-    await _repository.deleteMessages(removedMessages);
-
-    final updatedSession = session.copyWith(
-      messages: retainedMessages,
-      updatedAt: DateTime.now(),
-    );
-    final savedSession = await _saveSessionWithMessagesLocally(
-      updatedSession,
-      retainedMessages,
-      notify: false,
-    );
-    await _generateAssistantResponse(
-      savedSession,
-      model,
-      sessionMessages: retainedMessages,
-    );
-  }
-
-  Future<void> _regenerateAssistantMessageVersion({
-    required ChatSessionRecord session,
-    required List<ChatMessageRecord> messages,
-    required int targetIndex,
-    required ChatModelOption model,
-  }) async {
-    final targetMessage = messages[targetIndex];
-    final promptMessages = List<ChatMessageRecord>.from(
-      messages.take(targetIndex),
-    );
-    final retainedMessages = List<ChatMessageRecord>.from(
-      messages.take(targetIndex + 1),
-    );
-    final removedMessages = messages.skip(targetIndex + 1);
-    await _repository.deleteMessages(removedMessages);
-
-    var retainedTargetMessage = targetMessage;
-    var versionIds = List<String>.from(targetMessage.versionIds);
-    if (versionIds.isEmpty) {
-      final originalVersion = _versionFromMessage(targetMessage);
-      await _repository.saveMessageVersion(originalVersion);
-      versionIds = <String>[originalVersion.id];
-      retainedTargetMessage = targetMessage.copyWith(
-        versionIds: versionIds,
-        currentVersionIndex: 0,
-      );
-      retainedMessages[targetIndex] = retainedTargetMessage;
-    }
-
-    final fallbackSession = session.copyWith(
-      messages: retainedMessages,
-      updatedAt: DateTime.now(),
-    );
-    final savedFallbackSession = await _saveSessionWithMessagesLocally(
-      fallbackSession,
-      retainedMessages,
-      notify: false,
-    );
-
-    final now = DateTime.now();
-    final nextVersion = ChatMessageVersionRecord(
-      id: _generateId('version'),
-      messageId: targetMessage.id,
-      content: '',
-      createdAt: now,
-      modelName: model.displayName,
-      reasoningContent: '',
-    );
-    final nextVersionIds = <String>[...versionIds, nextVersion.id];
-    final nextVersionIndex = nextVersionIds.length - 1;
-    retainedTargetMessage = retainedTargetMessage.copyWith(
-      content: '',
-      createdAt: now,
-      modelName: model.displayName,
-      reasoningContent: '',
-      clearImageFilePaths: true,
-      versionIds: nextVersionIds,
-      currentVersionIndex: nextVersionIndex,
-    );
-    final streamingMessages = List<ChatMessageRecord>.from(retainedMessages);
-    streamingMessages[targetIndex] = retainedTargetMessage;
-    final streamingSession = savedFallbackSession.copyWith(
-      messages: streamingMessages,
-      updatedAt: now,
-    );
-
-    await _generateAssistantResponse(
-      streamingSession,
-      model,
-      sessionMessages: streamingMessages,
-      promptMessages: promptMessages,
-      draftSeed: retainedTargetMessage,
-      versionId: nextVersion.id,
-      emptyDraftFallbackSession: savedFallbackSession,
-      emptyDraftFallbackMessages: retainedMessages,
-    );
+    await _conversation.loadSelectedSessionMessages(sessionId);
   }
 
   Future<void> selectSession(String sessionId) async {
     if (!canManageSessions) {
       return;
     }
-    if (_selectedSessionId == sessionId && !_isLoadingMessages) {
-      return;
-    }
-    final session = _findSession(sessionId);
-    if (session == null) {
-      return;
-    }
-    final generation = ++_sessionLoadGeneration;
-    _selectedSessionId = sessionId;
-    _visibleMessages = <ChatMessageRecord>[];
-    _isLoadingMessages = true;
-    notifyListeners();
-
-    try {
-      final messages = await _repository.loadRecentMessages(
-        session,
-        limit: messageWindowSize,
-      );
-      if (generation != _sessionLoadGeneration) {
-        return;
-      }
-      _visibleMessages = messages;
-    } catch (_) {
-      if (generation != _sessionLoadGeneration) {
-        return;
-      }
-      _visibleMessages = <ChatMessageRecord>[];
-    } finally {
-      if (generation == _sessionLoadGeneration) {
-        _isLoadingMessages = false;
-        notifyListeners();
-      }
-    }
+    await _conversation.selectSession(sessionId);
   }
 
   Future<void> loadOlderMessages() async {
-    final session = selectedSession;
-    if (session == null ||
-        _visibleMessages.isEmpty ||
-        _isLoadingMessages ||
-        _isLoadingOlderMessages ||
-        !hasOlderMessages) {
-      return;
-    }
-
-    _isLoadingOlderMessages = true;
-    notifyListeners();
-
-    try {
-      final olderMessages = await _repository.loadMessagesBefore(
-        session,
-        beforeMessageId: _visibleMessages.first.id,
-        limit: messageWindowSize,
-      );
-      if (olderMessages.isNotEmpty && _selectedSessionId == session.id) {
-        _visibleMessages = <ChatMessageRecord>[
-          ...olderMessages,
-          ..._visibleMessages,
-        ];
-      }
-    } catch (_) {
-    } finally {
-      _isLoadingOlderMessages = false;
-      notifyListeners();
-    }
+    await _conversation.loadOlderMessages();
   }
 
   void selectLoadedModel(String modelId) {
     if (!canSelectModels) {
       return;
     }
-    final model = _findModel(modelId);
-    if (model == null || !model.isLoaded) {
-      return;
-    }
-    _currentModelId = modelId;
-    notifyListeners();
+    _models.selectLoadedModel(modelId);
   }
 
   Future<void> loadAndSelectModel(String modelId) async {
     if (!canSelectModels) {
       return;
     }
-
-    final model = _findModel(modelId);
-    if (model == null) {
-      return;
-    }
-
-    if (model.isLoaded) {
-      selectLoadedModel(modelId);
-      return;
-    }
-
-    _loadingModelId = modelId;
-    notifyListeners();
-
-    try {
-      await _apiClient.loadModel(modelId);
-      _models = await _apiClient.fetchModels();
-      _currentModelId = modelId;
-    } catch (_) {
-      try {
-        _models = await _apiClient.fetchModels();
-      } catch (_) {}
-    } finally {
-      _loadingModelId = null;
-      notifyListeners();
-    }
+    await _models.loadAndSelectModel(modelId);
   }
 
   Future<void> loadCurrentModel() async {
-    final modelId = _currentModelId;
-    if (modelId == null) {
-      return;
-    }
-    await loadAndSelectModel(modelId);
+    await _models.loadCurrentModel();
   }
 
   Future<void> unloadModel(String modelId) async {
     if (!canSelectModels) {
       return;
     }
-
-    final model = _findModel(modelId);
-    if (model == null || !model.isLoaded) {
-      return;
-    }
-
-    _loadingModelId = modelId;
-    notifyListeners();
-
-    try {
-      await _apiClient.unloadModel(modelId);
-      _models = await _apiClient.fetchModels();
-      if (_currentModelId == modelId) {
-        _currentModelId = null;
-      } else {
-        final selectedModel = currentModel;
-        if (_currentModelId != null &&
-            (selectedModel == null || !selectedModel.isLoaded)) {
-          _currentModelId = null;
-        }
-      }
-    } catch (_) {
-      try {
-        _models = await _apiClient.fetchModels();
-      } catch (_) {}
-    } finally {
-      _loadingModelId = null;
-      notifyListeners();
-    }
+    await _models.unloadModel(modelId);
   }
 
   Future<void> sendMessage(
     String text, {
     List<String>? imageAttachments,
   }) async {
-    if (!canSend) {
-      return;
-    }
+    await _generation.sendMessage(text, imageAttachments: imageAttachments);
+  }
 
-    final model = currentModel;
-    final normalizedText = text.trim();
-    if (model == null ||
-        (normalizedText.isEmpty &&
-            (imageAttachments == null || imageAttachments.isEmpty))) {
-      return;
-    }
+  bool canRegenerateMessage(String messageId) {
+    return _generation.canRegenerateMessage(messageId);
+  }
 
-    var session = selectedSession;
-    if (session == null) {
-      session = _createSessionRecord();
-      _selectedSessionId = session.id;
-      _visibleMessages = <ChatMessageRecord>[];
-    }
+  Future<void> editMessage({
+    required String messageId,
+    required String newContent,
+  }) async {
+    await _generation.editMessage(messageId: messageId, newContent: newContent);
+  }
 
-    _isSending = true;
-    // 清空附件图片预览
-    clearImageAttachments();
+  Future<void> deleteMessage(String messageId) async {
+    await _generation.deleteMessage(messageId);
+  }
 
-    final now = DateTime.now();
-    final userMessage = ChatMessageRecord(
-      id: _generateId('message'),
-      role: ChatRole.user,
-      content: normalizedText,
-      createdAt: now,
-      sessionId: session.id,
-      modelName: model.displayName,
-      imageFilePaths: imageAttachments ?? const [],
+  Future<void> selectMessageVersion({
+    required String messageId,
+    required int versionIndex,
+  }) async {
+    await _generation.selectMessageVersion(
+      messageId: messageId,
+      versionIndex: versionIndex,
     );
+  }
 
-    final messages = await _repository.loadAllMessages(session);
-    final sessionTitle = session.title == defaultSessionTitle
-        ? _deriveSessionTitle(normalizedText)
-        : session.title;
-    final updatedMessages = <ChatMessageRecord>[...messages, userMessage];
-    final updatedSession = session.copyWith(
-      title: sessionTitle,
-      messages: updatedMessages,
-      updatedAt: now,
-    );
-    final savedSession = await _saveSessionWithMessagesLocally(
-      updatedSession,
-      updatedMessages,
-      notify: false,
-    );
-    await _generateAssistantResponse(
-      savedSession,
-      model,
-      sessionMessages: updatedMessages,
-    );
+  Future<void> regenerateFromMessage(String messageId) async {
+    await _generation.regenerateFromMessage(messageId);
   }
 
   void clearLastError() {
-    _lastErrorMessage = null;
+    _generation.clearLastError();
   }
 
   void cancelStreaming() {
-    _activeCancelToken?.cancel('user canceled');
+    _generation.cancelStreaming();
   }
 
   void addImageAttachment(String filePath) {
-    if (_pendingImageAttachments.length >= 5) {
-      return;
-    }
-    _pendingImageAttachments = List<String>.from(_pendingImageAttachments)
-      ..add(filePath);
-    notifyListeners();
+    _generation.addImageAttachment(filePath);
   }
 
   void removeImageAttachment(int index) {
-    if (index < 0 || index >= _pendingImageAttachments.length) {
-      return;
-    }
-    _pendingImageAttachments = List<String>.from(_pendingImageAttachments)
-      ..removeAt(index);
-    notifyListeners();
+    _generation.removeImageAttachment(index);
   }
 
   void clearImageAttachments() {
-    if (_pendingImageAttachments.isEmpty) {
-      return;
-    }
-    _pendingImageAttachments = [];
+    _generation.clearImageAttachments();
+  }
+
+  void _notify() {
     notifyListeners();
   }
 
-  Future<void> _saveSessionLocally(
-    ChatSessionRecord session, {
-    bool notify = true,
-  }) async {
-    await _repository.saveSession(session);
-    _upsertSession(session);
-    if (notify) {
-      notifyListeners();
-    }
-  }
-
-  Future<ChatSessionRecord> _saveSessionWithMessagesLocally(
-    ChatSessionRecord session,
-    List<ChatMessageRecord> messages, {
-    bool notify = true,
-  }) async {
-    final savedSession = await _repository.saveSessionWithMessages(
-      session,
-      messages,
-    );
-    _upsertSession(savedSession);
-    if (_selectedSessionId == savedSession.id) {
-      _syncVisibleMessagesFromFullMessages(messages);
-    }
-    if (notify) {
-      notifyListeners();
-    }
-    return savedSession;
-  }
-
-  void _upsertSession(ChatSessionRecord session) {
-    final nextSessions = List<ChatSessionRecord>.from(_sessions);
-    final index = nextSessions.indexWhere((item) => item.id == session.id);
-    if (index >= 0) {
-      nextSessions[index] = session;
-    } else {
-      nextSessions.add(session);
-    }
-    nextSessions.sort(
-      (left, right) => right.updatedAt.compareTo(left.updatedAt),
-    );
-    _sessions = nextSessions;
-  }
-
-  void _syncVisibleMessagesFromFullMessages(
-    List<ChatMessageRecord> fullMessages,
-  ) {
-    if (fullMessages.isEmpty) {
-      _visibleMessages = <ChatMessageRecord>[];
-      return;
-    }
-    if (_visibleMessages.isNotEmpty) {
-      final firstVisibleId = _visibleMessages.first.id;
-      final firstVisibleIndex = fullMessages.indexWhere(
-        (message) => message.id == firstVisibleId,
-      );
-      if (firstVisibleIndex >= 0) {
-        _visibleMessages = fullMessages
-            .skip(firstVisibleIndex)
-            .toList(growable: false);
-        return;
-      }
-    }
-    final targetCount = min(
-      fullMessages.length,
-      max(messageWindowSize, _visibleMessages.length),
-    );
-    _visibleMessages = fullMessages
-        .skip(fullMessages.length - targetCount)
-        .toList(growable: false);
-  }
-
-  Future<void> _generateAssistantResponse(
-    ChatSessionRecord session,
-    ChatModelOption model, {
-    required List<ChatMessageRecord> sessionMessages,
-    List<ChatMessageRecord>? promptMessages,
-    ChatMessageRecord? draftSeed,
-    String? versionId,
-    ChatSessionRecord? emptyDraftFallbackSession,
-    List<ChatMessageRecord>? emptyDraftFallbackMessages,
-  }) async {
-    _isSending = true;
-
-    final draftMessage =
-        draftSeed ??
-        ChatMessageRecord(
-          id: _generateId('draft'),
-          role: ChatRole.assistant,
-          content: '',
-          createdAt: DateTime.now(),
-          sessionId: session.id,
-          modelName: model.displayName,
-          reasoningContent: '',
-        );
-    _draftAssistantMessage = draftMessage;
-    _activeCancelToken = CancelToken();
-
-    var currentMessages = _messagesWithStreamingDraft(
-      sessionMessages,
-      draftMessage,
-      appendIfMissing: draftSeed == null,
-    );
-    var currentSession = session.copyWith(
-      messages: currentMessages,
-      updatedAt: DateTime.now(),
-    );
-    await _persistStreamingDraft(
-      session: currentSession,
-      messages: currentMessages,
-      draft: draftMessage,
-      versionId: versionId,
-      notify: true,
-    );
-
-    try {
-      await for (final delta in _apiClient.streamChatCompletion(
-        modelId: model.id,
-        messages: promptMessages ?? sessionMessages,
-        cancelToken: _activeCancelToken!,
-      )) {
-        final currentDraft = _draftAssistantMessage;
-        if (currentDraft == null) {
-          continue;
-        }
-        final nextContent = delta.content.isEmpty
-            ? currentDraft.content
-            : '${currentDraft.content}${delta.content}';
-        final currentReasoningContent = currentDraft.reasoningContent ?? '';
-        final nextReasoningContent = delta.reasoningContent.isEmpty
-            ? currentDraft.reasoningContent
-            : '$currentReasoningContent${delta.reasoningContent}';
-        _draftAssistantMessage = currentDraft.copyWith(
-          content: nextContent,
-          reasoningContent: nextReasoningContent,
-        );
-        currentMessages = _messagesWithStreamingDraft(
-          currentMessages,
-          _draftAssistantMessage!,
-          appendIfMissing: draftSeed == null,
-        );
-        currentSession = currentSession.copyWith(
-          messages: currentMessages,
-          updatedAt: DateTime.now(),
-        );
-        await _persistStreamingDraft(
-          session: currentSession,
-          messages: currentMessages,
-          draft: _draftAssistantMessage!,
-          versionId: versionId,
-          notify: true,
-        );
-      }
-    } catch (error) {
-      if (error is DioException && CancelToken.isCancel(error)) {
-        // 用户主动取消，不生成错误消息
-      } else {
-        final draft = _draftAssistantMessage;
-        final errorMessage = _chatErrorMessage(error);
-        if (draft != null && draft.content.trim().isNotEmpty) {
-          _lastErrorMessage = errorMessage;
-        } else {
-          _draftAssistantMessage = ChatMessageRecord(
-            id: draft?.id ?? _generateId('error'),
-            role: ChatRole.assistant,
-            content: errorMessage,
-            createdAt: DateTime.now(),
-            sessionId: session.id,
-            modelName: currentModel?.displayName,
-            versionIds: draft?.versionIds ?? const [],
-            currentVersionIndex: draft?.currentVersionIndex ?? 0,
-          );
-          currentMessages = _messagesWithStreamingDraft(
-            currentMessages,
-            _draftAssistantMessage!,
-            appendIfMissing: draftSeed == null,
-          );
-          currentSession = currentSession.copyWith(
-            messages: currentMessages,
-            updatedAt: DateTime.now(),
-          );
-          await _persistStreamingDraft(
-            session: currentSession,
-            messages: currentMessages,
-            draft: _draftAssistantMessage!,
-            versionId: versionId,
-            notify: true,
-          );
-        }
-      }
-    } finally {
-      final draft = _draftAssistantMessage;
-      final hasDraftContent = draft != null && draft.content.trim().isNotEmpty;
-      final hasDraftReasoning =
-          draft != null && (draft.reasoningContent?.trim().isNotEmpty ?? false);
-      if (draft != null && !hasDraftContent && !hasDraftReasoning) {
-        if (versionId != null) {
-          await _repository.deleteMessageVersions(<String>[versionId]);
-          final fallbackSession = emptyDraftFallbackSession;
-          final fallbackMessages = emptyDraftFallbackMessages;
-          if (fallbackSession != null && fallbackMessages != null) {
-            await _saveSessionWithMessagesLocally(
-              fallbackSession,
-              fallbackMessages,
-              notify: false,
-            );
-          }
-        } else {
-          final cleanedMessages = _messagesWithoutMessage(
-            currentMessages,
-            draft.id,
-          );
-          await _repository.deleteMessages(<ChatMessageRecord>[draft]);
-          final cleanedSession = currentSession.copyWith(
-            messages: cleanedMessages,
-            updatedAt: DateTime.now(),
-          );
-          await _saveSessionWithMessagesLocally(
-            cleanedSession,
-            cleanedMessages,
-            notify: false,
-          );
-        }
-      }
-
-      _draftAssistantMessage = null;
-      _activeCancelToken = null;
-      _isSending = false;
-      _pendingImageAttachments = [];
-      notifyListeners();
-    }
-  }
-
-  Future<void> _persistStreamingDraft({
-    required ChatSessionRecord session,
-    required List<ChatMessageRecord> messages,
-    required ChatMessageRecord draft,
-    required String? versionId,
-    required bool notify,
-  }) async {
-    if (versionId != null) {
-      await _repository.saveMessageVersion(
-        ChatMessageVersionRecord(
-          id: versionId,
-          messageId: draft.id,
-          content: draft.content,
-          createdAt: draft.createdAt,
-          modelName: draft.modelName,
-          reasoningContent: draft.reasoningContent,
-          imageFilePaths: draft.imageFilePaths,
-        ),
-      );
-    }
-    await _saveSessionWithMessagesLocally(session, messages, notify: notify);
-  }
-
-  List<ChatMessageRecord> _messagesWithStreamingDraft(
-    List<ChatMessageRecord> messages,
-    ChatMessageRecord draft, {
-    required bool appendIfMissing,
-  }) {
-    final nextMessages = List<ChatMessageRecord>.from(messages);
-    final index = _messageIndex(nextMessages, draft.id);
-    if (index >= 0) {
-      nextMessages[index] = draft;
-    } else if (appendIfMissing) {
-      nextMessages.add(draft);
-    } else {
-      return messages;
-    }
-    return nextMessages;
-  }
-
-  List<ChatMessageRecord> _messagesWithoutMessage(
-    List<ChatMessageRecord> messages,
-    String messageId,
-  ) {
-    return List<ChatMessageRecord>.from(messages)
-      ..removeWhere((message) => message.id == messageId);
-  }
-
-  ChatSessionRecord _createSessionRecord() {
-    final now = DateTime.now();
-    return ChatSessionRecord(
-      id: _generateId('session'),
-      title: defaultSessionTitle,
-      messageIds: const <String>[],
-      createdAt: now,
-      updatedAt: now,
-    );
-  }
-
-  ChatSessionRecord? _findSession(String? sessionId) {
-    if (sessionId == null) {
-      return null;
-    }
-    for (final session in _sessions) {
-      if (session.id == sessionId) {
-        return session;
-      }
-    }
-    return null;
-  }
-
-  ChatModelOption? _findModel(String? modelId) {
-    if (modelId == null) {
-      return null;
-    }
-    for (final model in _models) {
-      if (model.id == modelId) {
-        return model;
-      }
-    }
-    return null;
-  }
-
-  int _messageIndex(List<ChatMessageRecord> messages, String messageId) {
-    return messages.indexWhere((message) => message.id == messageId);
-  }
-
-  int _nearestUserMessageIndex(
-    List<ChatMessageRecord> messages,
-    int startIndex,
-  ) {
-    for (var index = startIndex; index >= 0; index -= 1) {
-      if (messages[index].role == ChatRole.user) {
-        return index;
-      }
-    }
-    return -1;
-  }
-
-  String _deriveSessionTitle(String text) {
-    final normalized = text.replaceAll(RegExp(r'\s+'), ' ').trim();
-    if (normalized.isEmpty) {
-      return defaultSessionTitle;
-    }
-    if (normalized.length <= 18) {
-      return normalized;
-    }
-    return '${normalized.substring(0, 18)}...';
-  }
-
-  String _generateId(String prefix) {
-    return '${prefix}_${DateTime.now().microsecondsSinceEpoch}_${_random.nextInt(1 << 32)}';
-  }
-
-  String? _selectedVersionId(ChatMessageRecord message) {
-    if (message.versionIds.isEmpty ||
-        message.currentVersionIndex < 0 ||
-        message.currentVersionIndex >= message.versionIds.length) {
-      return null;
-    }
-    return message.versionIds[message.currentVersionIndex];
-  }
-
-  ChatMessageVersionRecord _versionFromMessage(
-    ChatMessageRecord message, {
-    String? messageId,
-  }) {
-    final targetMessageId = messageId ?? message.id;
-    return ChatMessageVersionRecord(
-      id: _generateId('version'),
-      messageId: targetMessageId,
-      content: message.content,
-      createdAt: message.createdAt,
-      modelName: message.modelName,
-      reasoningContent: message.reasoningContent,
-      imageFilePaths: message.imageFilePaths,
-    );
-  }
-
-  ChatMessageRecord _messageWithVersion(
-    ChatMessageRecord message,
-    ChatMessageVersionRecord version,
-    int versionIndex,
-  ) {
-    return message.copyWith(
-      content: version.content,
-      createdAt: version.createdAt,
-      modelName: version.modelName,
-      clearModelName: version.modelName == null,
-      reasoningContent: version.reasoningContent,
-      clearReasoningContent: version.reasoningContent == null,
-      imageFilePaths: version.imageFilePaths,
-      currentVersionIndex: versionIndex,
-    );
-  }
-
-  String get _normalizedSessionQuery => _sessionQuery.trim().toLowerCase();
-
-  String _chatErrorMessage(Object error) {
-    return "Request Failed: $error";
-  }
-
-  String _modelStatusLabel(ChatModelStatus status) {
-    switch (status) {
-      case ChatModelStatus.loaded:
-        return '已加载';
-      case ChatModelStatus.loading:
-        return '加载中';
-      case ChatModelStatus.unloaded:
-        return '可加载';
-      case ChatModelStatus.failed:
-        return '加载失败';
-    }
+  @override
+  void dispose() {
+    _sessionList.removeListener(_notify);
+    _conversation.removeListener(_notify);
+    _models.removeListener(_notify);
+    _generation.removeListener(_notify);
+    _generation.dispose();
+    _models.dispose();
+    _conversation.dispose();
+    _sessionList.dispose();
+    super.dispose();
   }
 }
